ENH: Prep and Seg modules enhancements
Prep: correcte the bug of brainstem and cerebellum loading files
Seg: General segmetnation improved in order to use global enhancement module and also minor errors of tissues in wrong places. The CSF, GM and WM tissues now present better delineation. 
The median filter was set as optional...however, it may be interesting put it as fixed.