ENH: BVR is faster
The skeleton method was removed and the contour is updated on each iteration.
The iterations are only applied over the smallest regions that contains the brain contour. This reduces significantly the total processing time.
The median filter is still needed for stable results, which need further evaluation.
The total time processing was added.