STY: Compiling organization and removing unnecessary folders
The resources folder in Preparation module was removed since the Segmentation resources folder is used instead.
The CMakeList of Segmentation module was edited in order to add the processing scripts for both Preparation and Segmentation modules. Now the user can add those scripts in the system PATH and process those python modules easily.